Adds watch, unwatch, and user messaging
The bot now detects when a user mentions it
in a comment. Based off the text in the user comment
determines the action taken.
When the followed posts are updated, the users will be notified
in their inbox with the submission that was changed and what exactly
was changed via difftool

@BOT_NAME_HERE watch
- Comment this in any submission and the bot will follow this post for
you, notifying you about any changes

@BOT_NAME_HERE unwatch
- Comment this in the submission you are currently watching to no longer
be notified on changes for it